+++ include/spead2/recv_chunk_stream.h
@@ -97,12 +97,12 @@ typedef std::function<void(chunk_place_data *data, std::size_t data_size)> chunk
  */
 // If this is updated, update doc/cpp-recv-chunk.rst as well. It's not
 // documented with breathe due to a Doxygen bug.
-typedef std::function<std::unique_ptr<chunk>(std::int64_t chunk_id)> chunk_allocate_function;
+typedef std::function<std::unique_ptr<chunk>(std::int64_t chunk_id, std::uint64_t *batch_stats)> chunk_allocate_function;
 
 /**
  * Callback to receive a completed chunk. It takes ownership of the chunk.
  */
-typedef std::function<void(std::unique_ptr<chunk> &&)> chunk_ready_function;
+typedef std::function<void(std::unique_ptr<chunk> &&, std::uint64_t *batch_stats)> chunk_ready_function;
 
 /**
  * Parameters for a @ref chunk_stream.
@@ -359,10 +359,16 @@ public:
     /**
      * Constructor. Refer to @ref chunk_stream::chunk_stream for details.
      *
-     * The @link chunk_stream_config::set_allocate allocate@endlink and
-     * @link chunk_stream_config::set_ready ready@endlink callbacks are
-     * ignored and should be unset. Calling @ref get_chunk_config will
-     * reflect the internally-defined callbacks.
+     * The @link chunk_stream_config::set_allocate allocate@endlink callback
+     * is ignored and should be unset. If a
+     * @link chunk_stream_config::set_ready ready@endlink callback is
+     * defined, it will be called before the chunk is pushed onto the
+     * ringbuffer. It must not move from the provided @a unique_ptr, but it
+     * can be used to perform further processing on the chunk before it is
+     * pushed.
+     *
+     * Calling @ref get_chunk_config will reflect the internally-defined
+     * callbacks.
      */
     chunk_ring_stream(
         io_service_ref io_service,
@@ -420,7 +426,7 @@ chunk_stream_config chunk_ring_stream<DataRingbuffer, FreeRingbuffer>::adjust_ch
 {
     chunk_stream_config new_config = chunk_config;
     // Set the allocate callback to get a chunk from the free ringbuffer
-    new_config.set_allocate([&free_ring] (std::int64_t) -> std::unique_ptr<chunk> {
+    new_config.set_allocate([&free_ring] (std::int64_t, std::uint64_t *) -> std::unique_ptr<chunk> {
         try
         {
             return free_ring.pop();
@@ -433,9 +439,14 @@ chunk_stream_config chunk_ring_stream<DataRingbuffer, FreeRingbuffer>::adjust_ch
         }
     });
     // Set the ready callback to push chunks to the data ringbuffer
-    new_config.set_ready([&data_ring, &graveyard] (std::unique_ptr<chunk> &&c) {
+    auto orig_ready = chunk_config.get_ready();
+    new_config.set_ready(
+        [&data_ring, &graveyard, orig_ready] (std::unique_ptr<chunk> &&c,
+                                              std::uint64_t *batch_stats) {
         try
         {
+            if (orig_ready)
+                orig_ready(std::move(c), batch_stats);
             // TODO: use try_push and track stalls
             data_ring.push(std::move(c));
         }