+++ manylinux/before_all.sh
@@ -7,19 +7,19 @@ set -e -u
 package="$1"
 
 yum install -y \
-    wget libpcap libpcap-devel python-devel \
+    libpcap libpcap-devel python-devel \
     cmake3 ninja-build pandoc libnl3-devel \
     ccache
 
 # Install boost
-wget --progress=dot:mega https://ufpr.dl.sourceforge.net/project/boost/boost/1.74.0/boost_1_74_0.tar.bz2 -O /tmp/boost_1_74_0.tar.bz2
+curl -L https://ufpr.dl.sourceforge.net/project/boost/boost/1.74.0/boost_1_74_0.tar.bz2 -o /tmp/boost_1_74_0.tar.bz2
 tar -C /tmp -jxf /tmp/boost_1_74_0.tar.bz2
 cd /tmp/boost_1_74_0
 ./bootstrap.sh --prefix=/usr --with-libraries=program_options,system
 ./b2 cxxflags=-fPIC link=static install
 
 # Install rdma-core
-wget --progress=dot:mega https://github.com/linux-rdma/rdma-core/releases/download/v31.0/rdma-core-31.0.tar.gz -O /tmp/rdma-core-31.0.tar.gz
+curl -L https://github.com/linux-rdma/rdma-core/releases/download/v31.0/rdma-core-31.0.tar.gz -o /tmp/rdma-core-31.0.tar.gz
 tar -C /tmp -zxf /tmp/rdma-core-31.0.tar.gz
 cd /tmp/rdma-core-31.0
 mkdir build